Change Deeltaak.voorgaandeTaak to List<Deeltaak>
As part of the fix to the doorlooptijd calculation function we need to
be able to keep track of potentially multiple preceding tasks. To make
this possible voorgaandeTaak needs to be a list of deeltaken.

Also made a small change to getTakenduur in order to make this compile.
This function will be replaced anyway.